fix/frontend: 관광지 후기 삭제 시 연관 데이터 함께 삭제
관광지 후기 삭제 시 관광지 후기 공감 데이터도 함께 삭제되어야함

+++ backend/src/main/java/com/positivewand/tourin/domain/tourspot/TourSpotReviewService.java
@@ -68,6 +68,13 @@ public class TourSpotReviewService {
 
     @Transactional
     public void deleteTourSpotReview(Long tourSpotReviewId) {
+        Optional<TourSpotReview> tourSpotReview = tourSpotReviewRepository.findById(tourSpotReviewId);
+
+        if(tourSpotReview.isEmpty()) {
+            throw new NoSuchElementException("관광지 후기가 존재하지 않습니다.");
+        }
+
+        tourSpotReviewLikeRepository.deleteByTourSpotReviewId(tourSpotReviewId);
         tourSpotReviewRepository.deleteById(tourSpotReviewId);
     }
 }
